web-scrapper-service/requirements.txt
fastapi==0.79.0
uvicorn==0.17.6
httpx==0.24.0
beautifulsoup4==4.11.1
selenium~=4.16.0
webdriver_manager
celery~=5.3.6
requests~=2.31.0
pydantic~=1.10.13
python-dotenv~=1.0.0
flower~=2.0.1
redis~=5.0.1
